Add qa.ts
