+++ src/devices/tuya.ts
@@ -6073,6 +6073,7 @@ const definitions: Definition[] = [
         exposes: [ez.learn_ir_code(), ez.learned_ir_code(), ez.ir_code_to_send()],
         whiteLabel: [
             tuya.whitelabel('Tuya', 'UFO-R4Z', 'Universal smart IR remote control', ['_TZ3290_rlkmy85q4pzoxobl']),
+            tuya.whitelabel('QA', 'QAIRZPRO', 'Infrared hub pro', ['_TZ3290_jxvzqatwgsaqzx1u']),
         ],
     },
     {
@@ -8702,398 +8703,10 @@ const definitions: Definition[] = [
         },
     },
     {
-        fingerprint: tuya.fingerprint('TS0726', ['_TZ3000_kt6xxa4o']),
-        model: 'QAT42Z3',
-        vendor: 'QA',
-        description: 'QA 3CH Scene Switch',
-        whiteLabel: [{vendor: 'QA', model: 'QAT42ZT3'}],
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1, 'l2': 2, 'l3': 3}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1', 'l2', 'l3'], powerOnBehavior2: true, backlightModeOffOn: true}),
-            actionEnumLookup({
-                cluster: 'genOnOff',
-                commands: ['commandTuyaAction'],
-                attribute: 'value',
-                actionLookup: {'button': 0},
-                buttonLookup: {
-                    '1_up': 4, '1_down': 1,
-                    '2_up': 5, '2_down': 2,
-                    '3_up': 6, '3_down': 3,
-                },
-            }),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0726', ['_TZ3000_wopf2sox']),
-        model: 'QAT42Z1',
-        vendor: 'QA',
-        description: 'QA 1CH Scene Switch',
-        whiteLabel: [{vendor: 'QA', model: 'QAT42ZT1'}],
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1'], powerOnBehavior2: true, backlightModeOffOn: true}),
-            actionEnumLookup({
-                cluster: 'genOnOff',
-                commands: ['commandTuyaAction'],
-                attribute: 'value',
-                actionLookup: {'button': 0},
-                buttonLookup: {
-                    '1_up': 4, '1_down': 1,
-                    '2_up': 5, '2_down': 2,
-                    '3_up': 6, '3_down': 3,
-                },
-            }),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0726', ['_TZ3000_ssup6h68']),
-        model: 'QAT42Z2',
-        vendor: 'QA',
-        description: 'QA 2CH Scene Switch',
-        whiteLabel: [{vendor: 'QA', model: 'QAT42ZT2'}],
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1, 'l2': 2}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1', 'l2'], powerOnBehavior2: true, backlightModeOffOn: true}),
-            actionEnumLookup({
-                cluster: 'genOnOff',
-                commands: ['commandTuyaAction'],
-                attribute: 'value',
-                actionLookup: {'button': 0},
-                buttonLookup: {
-                    '1_up': 4, '1_down': 1,
-                    '2_up': 5, '2_down': 2,
-                    '3_up': 6, '3_down': 3,
-                },
-            }),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0001', ['_TZ3000_gtdswg8k']),
-        model: 'QARZDC1LR',
-        vendor: 'QA',
-        description: 'QA 1CH Long Range',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1'], powerOnBehavior2: true, switchType: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0002', ['_TZ3000_rfjctviq', '_TZ3210_pdnwpnz5', '_TZ3210_a2erlvb8']),
-        model: 'QARZ2LR',
-        vendor: 'QA',
-        description: 'QA 2CH Long Range',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1, 'l2': 2}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1', 'l2'], powerOnBehavior2: true, switchType: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0003', ['_TZ3000_zeuulson']),
-        model: 'QARZ3LR',
-        vendor: 'QA',
-        description: 'QA 3CH Long Range',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1, 'l2': 2, 'l3': 3}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1', 'l2', 'l3'], powerOnBehavior2: true, switchType: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0004', ['_TZ3000_wwtnshol']),
-        model: 'QARZ4LR',
-        vendor: 'QA',
-        description: 'QA 4CH Long Range',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1, 'l2': 2, 'l3': 3, 'l4': 4}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1', 'l2', 'l3', 'l4'], powerOnBehavior2: true, switchType: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS1201', ['_TZ3290_jxvzqatwgsaqzx1u']),
-        model: 'QAIRZPRO',
-        vendor: 'QA',
-        description: 'QA IR Hub',
-        whiteLabel: [{vendor: 'QA', model: 'QAIRZPRO'}],
-        fromZigbee: [
-            zosung.fzZosung.zosung_send_ir_code_00, zosung.fzZosung.zosung_send_ir_code_01, zosung.fzZosung.zosung_send_ir_code_02,
-            zosung.fzZosung.zosung_send_ir_code_03, zosung.fzZosung.zosung_send_ir_code_04, zosung.fzZosung.zosung_send_ir_code_05,
-        ],
-        toZigbee: [zosung.tzZosung.zosung_ir_code_to_send, zosung.tzZosung.zosung_learn_ir_code],
-        exposes: [zosung.presetsZosung.learn_ir_code(), zosung.presetsZosung.learned_ir_code(), zosung.presetsZosung.ir_code_to_send()],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0001', [
-            '_TZ3000_dov0a3p1',
-        ]),
-        model: 'QAT42Z1H',
-        vendor: 'QA',
-        description: 'QA Switch 1CH',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1'], powerOnBehavior2: true, backlightModeOffOn: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0002', [
-            '_TZ3000_gkesadus',
-        ]),
-        model: 'QAT42Z2H',
-        vendor: 'QA',
-        description: 'QA Switch 2CH',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'l1': 1, 'l2': 2}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['l1', 'l2'], powerOnBehavior2: true, backlightModeOffOn: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0003', [
-            '_TZ3000_pmsxmttq',
-        ]),
-        model: 'QAT42Z3H',
-        vendor: 'QA',
-        description: 'QA Switch 3CH',
-        extend: [
-            tuya.modernExtend.tuyaMagicPacket(),
-            deviceEndpoints({endpoints: {'left': 1, 'center': 2, 'right': 3}}),
-            tuya.modernExtend.tuyaOnOff({endpoints: ['left', 'center', 'right'], powerOnBehavior2: true, backlightModeOffOn: true}),
-        ],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0601', [
-            '_TZE204_4cl0dzt4',
-        ]),
-        model: 'QAT44Z6H',
-        vendor: 'QA',
-        description: 'QA Switch 4x4 6CH',
-        exposes: [e.switch().withEndpoint('l1').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l2').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l3').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l4').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l5').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l6').setAccess('state', ea.STATE_SET)],
-        fromZigbee: [fz.ignore_basic_report, legacy.fz.tuya_switch],
-        toZigbee: [legacy.tz.tuya_switch_state],
-        meta: {multiEndpoint: true},
-        endpoint: (device) => {
-            return {'l1': 1, 'l2': 1, 'l3': 1, 'l4': 1, 'l5': 1, 'l6': 1};
-        },
-        configure: async (device, coordinatorEndpoint, logger) => {
-            await reporting.bind(device.getEndpoint(1), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(2)) await reporting.bind(device.getEndpoint(2), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(3)) await reporting.bind(device.getEndpoint(3), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(4)) await reporting.bind(device.getEndpoint(4), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(5)) await reporting.bind(device.getEndpoint(5), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(6)) await reporting.bind(device.getEndpoint(6), coordinatorEndpoint, ['genOnOff']);
-            // Reports itself as battery which is not correct: https://github.com/Koenkk/zigbee2mqtt/issues/6190
-            device.powerSource = 'Mains (single phase)';
-            device.save();
-        },
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0601', [
-            '_TZE204_kyzjsjo3',
-        ]),
-        model: 'QAT44Z4H',
-        vendor: 'QA',
-        description: 'QA Switch 4x4 4CH',
-        exposes: [e.switch().withEndpoint('l1').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l2').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l3').setAccess('state', ea.STATE_SET),
-            e.switch().withEndpoint('l4').setAccess('state', ea.STATE_SET)],
-        fromZigbee: [fz.ignore_basic_report, legacy.fz.tuya_switch],
-        toZigbee: [legacy.tz.tuya_switch_state],
-        meta: {multiEndpoint: true},
-        endpoint: (device) => {
-            return {'l1': 1, 'l2': 1, 'l3': 1, 'l4': 1};
-        },
-        configure: async (device, coordinatorEndpoint, logger) => {
-            await reporting.bind(device.getEndpoint(1), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(2)) await reporting.bind(device.getEndpoint(2), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(3)) await reporting.bind(device.getEndpoint(3), coordinatorEndpoint, ['genOnOff']);
-            if (device.getEndpoint(4)) await reporting.bind(device.getEndpoint(4), coordinatorEndpoint, ['genOnOff']);
-            // Reports itself as battery which is not correct: https://github.com/Koenkk/zigbee2mqtt/issues/6190
-            device.powerSource = 'Mains (single phase)';
-            device.save();
-        },
-    },
-    {
-        fingerprint: tuya.fingerprint('TS130F', [
-            '_TZ3210_xbpt8ewc',
-        ]),
-        model: 'QACZ1',
-        vendor: 'QA',
-        description: 'QA Curtain switch',
-        fromZigbee: [fz.tuya_cover_options, fz.cover_position_tilt],
-        toZigbee: [tz.cover_state, tz.moes_cover_calibration, tz.cover_position_tilt, tz.tuya_cover_reversal],
-        exposes: [e.cover_position(), e.numeric('calibration_time', ea.ALL).withValueMin(0).withValueMax(100),
-            e.enum('moving', ea.STATE, ['UP', 'STOP', 'DOWN']), e.binary('motor_reversal', ea.ALL, 'ON', 'OFF')],
-    },
-    {
-        fingerprint: tuya.fingerprint('TS0601', ['_TZE204_4qznlkbu']),
-        model: 'QASZ24R',
-        vendor: 'QA',
-        description: 'QA mmWave 24Ghz relay 10A',
-        configure: tuya.configureMagicPacket,
-        fromZigbee: [tuya.fz.datapoints],
-        toZigbee: [tuya.tz.datapoints],
-        exposes: [
-            e.presence(),
-            e.illuminance().withUnit('lx'),
-            e
-                .numeric('target_distance', ea.STATE)
-                .withDescription('Distance to target')
-                .withUnit('m'),
-            e
-                .numeric('radar_sensitivity', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(9)
-                .withValueStep(1)
-                .withDescription('Sensitivity of the radar'),
-            e
-                .numeric('entry_sensitivity', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(9)
-                .withValueStep(1)
-                .withDescription('Entry sensitivity'),
-            e
-                .numeric('detection_range', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(8)
-                .withValueStep(0.1)
-                .withUnit('m')
-                .withDescription('Detection range'),
-            e
-                .numeric('shield_range', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(8)
-                .withValueStep(0.1)
-                .withUnit('m')
-                .withDescription('Shield range of the radar'),
-            e
-                .numeric('entry_distance_indentation', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(8)
-                .withValueStep(0.1)
-                .withUnit('m')
-                .withDescription('Entry distance indentation'),
-            e
-                .numeric('entry_filter_time', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(10)
-                .withValueStep(0.1)
-                .withUnit('s')
-                .withDescription('Entry filter time'),
-            e
-                .numeric('departure_delay', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(600)
-                .withValueStep(1)
-                .withUnit('s')
-                .withDescription('Turn off delay'),
-            e
-                .numeric('block_time', ea.STATE_SET)
-                .withValueMin(0)
-                .withValueMax(10)
-                .withValueStep(0.1)
-                .withUnit('s')
-                .withDescription('Block time'),
-            e
-                .binary('breaker_status', ea.STATE_SET, 'ON', 'OFF')
-                .withDescription('Breaker status changes with breaker_mode->standard'),
-            e
-                .enum('breaker_mode', ea.STATE_SET, ['standard', 'local'])
-                .withDescription(
-                    'Status breaker mode: standard is external, local is auto'
-                ),
-            e
-                .enum('status_indication', ea.STATE_SET, ['OFF', 'ON'])
-                .withDescription('Led backlight when triggered'),
-        ],
-        meta: {
-            tuyaDatapoints: [
-                [1, 'presence', tuya.valueConverter.trueFalse1],
-                [2, 'radar_sensitivity', tuya.valueConverter.raw],
-                [3, 'shield_range', tuya.valueConverter.divideBy100],
-                [4, 'detection_range', tuya.valueConverter.divideBy100],
-                [6, 'equipment_status', tuya.valueConverter.raw],
-                [9, 'target_distance', tuya.valueConverter.divideBy100],
-                [101, 'entry_filter_time', tuya.valueConverter.divideBy10],
-                [102, 'departure_delay', tuya.valueConverter.raw],
-                [103, 'cline', tuya.valueConverter.raw],
-                [104, 'illuminance', tuya.valueConverter.divideBy10],
-                [105, 'entry_sensitivity', tuya.valueConverter.raw],
-                [106, 'entry_distance_indentation', tuya.valueConverter.divideBy100],
-                [
-                    107,
-                    'breaker_mode',
-                    tuya.valueConverterBasic.lookup({
-                        standard: tuya.enum(0),
-                        local: tuya.enum(1),
-                    }),
-                ],
-                [
-                    108,
-                    'breaker_status',
-                    tuya.valueConverterBasic.lookup({
-                        OFF: tuya.enum(0),
-                        ON: tuya.enum(1),
-                    }),
-                ],
-                [
-                    109,
-                    'status_indication',
-                    tuya.valueConverterBasic.lookup({
-                        OFF: tuya.enum(0),
-                        ON: tuya.enum(1),
-                    }),
-                ],
-                [
-                    111,
-                    'breaker_polarity',
-                    tuya.valueConverterBasic.lookup({
-                        NC: tuya.enum(0),
-                        NO: tuya.enum(1),
-                    }),
-                ],
-                [112, 'block_time', tuya.valueConverter.divideBy10],
-                [113, 'parameter_setting_result', tuya.valueConverter.raw],
-                [114, 'factory_parameters', tuya.valueConverter.raw],
-                [115, 'sensor', tuya.valueConverter.onOff],
-            ],
-        },
-    },
-    {
-        fingerprint: [{modelID: 'TS0203', manufacturerName: '_TZ3000_udyjylt7'},
-            {modelID: 'TS0203', manufacturerName: '_TZ3000_lltemgsf'},
-            {modelID: 'TS0203', manufacturerName: '_TZ3000_mg4dy6z6'}],
-        model: 'QASD1',
-        vendor: 'QA',
-        description: 'QA Door sensor',
-        fromZigbee: [fz.ias_contact_alarm_1, fz.battery, fz.ignore_basic_report, fz.ias_contact_alarm_1_report],
-        toZigbee: [],
-        exposes: [e.contact(), e.battery_low(), e.battery(), e.battery_voltage(), e.tamper(), e.linkquality()],
-        configure: async (device, coordinatorEndpoint) => {
-            try {
-                const endpoint = device.getEndpoint(1);
-                await reporting.bind(endpoint, coordinatorEndpoint, ['genPowerCfg']);
-                await reporting.batteryPercentageRemaining(endpoint);
-                await reporting.batteryVoltage(endpoint);
-            } catch (error) {/* Fails for some*/}
-        },
-    },
-    {
         fingerprint: tuya.fingerprint('TS110E', ['_TZ3210_zxbtub8r', '_TZ3210_k1msuvg6', '_TZ3210_hzdhb62z', '_TZ3210_v5yquxma']),
         model: 'QADZ1',
         vendor: 'QA',
-        description: 'QA Dimmer 1CH',
+        description: '1 channel dimmer',
         extend: [light({powerOnBehavior: false, configureReporting: true, effect: false})],
         fromZigbee: [tuya.fz.power_on_behavior_1, fzLocal.TS110E_switch_type, fzLocal.TS110E, fz.on_off],
         toZigbee: [tzLocal.TS110E_light_onoff_brightness, tuya.tz.power_on_behavior_1, tzLocal.TS110E_options],
@@ -9104,7 +8717,7 @@ const definitions: Definition[] = [
         fingerprint: tuya.fingerprint('TS110E', ['_TZ3210_tkkb1ym8']),
         model: 'QADZ2',
         vendor: 'QA',
-        description: 'QA Dimmer 2CH',
+        description: '2 channel dimmer',
         extend: [
             tuya.modernExtend.tuyaMagicPacket(),
             deviceEndpoints({endpoints: {'l1': 1, 'l2': 2}}),
